requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27

langchain==0.1.16
langchain-community==0.0.36
langchain-openai==0.1.7

openai==1.30.1
faiss-cpu==1.8.0
pypdf==4.1.0
python-dotenv==1.0.1
pandas==2.2.1
